Change IInstance GetDepth return type to double

+++ CraftingRPG/Interfaces/IInstance.cs
@@ -7,7 +7,7 @@ public interface IInstance
 {
     public Vector2 GetPosition();
     public Vector2 SetPosition(Vector2 position);
-    public float GetDepth();
+    public double GetDepth();
     public Vector2 GetSize();
     public int GetSpriteSheetIndex();
     public RectangleF GetCollisionBox();
